+++ libs/metaHDL_core/source/metaHDL_core/frontend/Bit.h
@@ -25,7 +25,7 @@ class Bit : public ElementarySignal
         
         Bit &operator=(const Bit &rhs) { assign(rhs); return *this; }
     protected:
-        virtual hlim::ConnectionType getSignalType(unsigned width) const override;
+        virtual hlim::ConnectionType getSignalType(size_t width) const override;
 };
 
 }
