+++ libs/metaHDL_core/source/metaHDL_core/hlim/ConnectionType.cpp
@@ -19,9 +19,9 @@ bool ConnectionType::operator==(const ConnectionType &rhs) const
     
 #if 0
 
-unsigned CompoundConnectionType::getTotalWidth() const 
+size_t CompoundConnectionType::getTotalWidth() const 
 {
-    unsigned sum = 0;
+    size_t sum = 0;
     for (const auto &sub : m_subConnections)
         sum += sub.type->getTotalWidth();
     return sum;
