+++ libs/metaHDL_core/source/metaHDL_core/hlim/coreNodes/Node_Signal.h
@@ -14,8 +14,8 @@ class Node_Signal : public Node
         
         virtual std::string getTypeName() const override { return "Signal"; }
         virtual void assertValidity() const override { }
-        virtual std::string getInputName(unsigned idx) const override { return "in"; }
-        virtual std::string getOutputName(unsigned idx) const override { return "out"; }
+        virtual std::string getInputName(size_t idx) const override { return "in"; }
+        virtual std::string getOutputName(size_t idx) const override { return "out"; }
         
         void setConnectionType(const ConnectionType &connectionType);
         inline const ConnectionType &getConnectionType() const { return m_connectionType; }
